# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /profile/me:
    get:
      operationId: getMyProfile
      tags: [profiles]
      summary: Get the current user's profile
      responses:
        "200":
          description: Profile found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Profile"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Profile not yet created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    put:
      operationId: upsertMyProfile
      tags: [profiles]
      summary: Create or update the current user's profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProfileInput"
      responses:
        "200":
          description: Profile saved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Profile"
        "400":
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /profiles:
    get:
      operationId: listProfiles
      tags: [profiles]
      summary: List all user profiles (admin)
      responses:
        "200":
          description: List of profiles
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Profile"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  schemas:
...
    Profile:
      type: object
      required: [id, nickname, age, email, memberType, interests, createdAt, updatedAt]
      properties:
        id:
          type: integer
        nickname:
          type: string
        age:
          type: integer
        email:
          type: string
        telegram:
          type: string
          nullable: true
        whatsapp:
          type: string
          nullable: true
        memberType:
          type: string
          enum: [singolo, coppia, singola, trav]
        interests:
          type: array
          items:
            type: string
            enum: [swinger, sexpositive, kinky, gangbang]
        createdAt:
          type: string
        updatedAt:
          type: string
    ProfileInput:
      type: object
      required: [nickname, age, email, memberType, interests]
      properties:
        nickname:
          type: string
          minLength: 1
        age:
          type: integer
          minimum: 18
        email:
          type: string
          minLength: 3
        telegram:
          type: string
        whatsapp:
          type: string
        memberType:
          type: string
          enum: [singolo, coppia, singola, trav]
        interests:
          type: array
          items:
            type: string
            enum: [swinger, sexpositive, kinky, gangbang]
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string