# lib/api-zod/src/generated/types/profile.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ProfileInterestsItem } from './profileInterestsItem';
import type { ProfileMemberType } from './profileMemberType';

export interface Profile {
  id: number;
  nickname: string;
  age: number;
  email: string;
  telegram?: string | null;
  whatsapp?: string | null;
  memberType: ProfileMemberType;
  interests: ProfileInterestsItem[];
  createdAt: string;
  updatedAt: string;
}