# lib/api-zod/src/generated/types/profileInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ProfileInputInterestsItem } from './profileInputInterestsItem';
import type { ProfileInputMemberType } from './profileInputMemberType';

export interface ProfileInput {
  /** @minLength 1 */
  nickname: string;
  /** @minimum 18 */
  age: number;
  /** @minLength 3 */
  email: string;
  telegram?: string;
  whatsapp?: string;
  memberType: ProfileInputMemberType;
  interests: ProfileInputInterestsItem[];
}